# docs/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "EvalAI API",
		"version": "1.0.0",
		"description": "AI Evaluation Platform API. See docs/api-contract.md for stability commitments."
	},
	"servers": [
		{
			"url": "https://v0-ai-evaluation-platform-nu.vercel.app",
			"description": "Production"
		},
		{
			"url": "http://localhost:3000",
			"description": "Local"
		}
	],
	"paths": {
...
		"/api/workflows/{id}/runs/{runId}": {
			"get": {
				"summary": "by id",
				"responses": {
					"200": {
						"description": "Success"
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer"
						}
					},
					{
						"name": "runId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer"
						}
					}
				]
			}
		},
...
	},
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "API Key or Session Token"
			}
		},
		"schemas": {
			"ApiError": {
				"type": "object",
				"properties": {
					"error": {
						"type": "object",
						"properties": {
							"code": {
								"type": "string",
								"enum": [
									"UNAUTHORIZED",
									"FORBIDDEN",
									"NOT_FOUND",
									"VALIDATION_ERROR",
									"RATE_LIMITED",
									"CONFLICT",
									"INTERNAL_ERROR",
									"SERVICE_UNAVAILABLE",
									"QUOTA_EXCEEDED",
									"NO_ORG_MEMBERSHIP"
								]
							},
							"message": {
								"type": "string"
							},
							"details": {},
							"requestId": {
								"type": "string",
								"format": "uuid"
							}
						}
					}
				}
			}
		},
		"responses": {
...
			"Unauthorized": {
				"description": "Unauthorized",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/ApiError"
						}
					}
				}
			}
		}
	}
}